+++ Marathon.Domain/Entities/RaceKitItem.cs
@@ -3,16 +3,16 @@ using Marathon.Domain.Common;
 
 namespace Marathon.Domain.Entities
 {
-    public sealed class RaceKitItem : IEntity<short>
+    public sealed class RaceKitItem : IEntity<ushort>
     {
         public RaceKitItem()
         {
             OptionItems = new HashSet<RaceKitOptionItem>();
         }
 
-        public short Id { get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
-        public long Count { get; set; }
+        public uint Count { get; set; }
 
 
         public ICollection<RaceKitOptionItem> OptionItems { get; set; }
